requirements.txt
click==8.1.3
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug>=2.3.7
zipp==3.10.0
openai==1.3.7
flask==2.3.3
flask_cors==4.0.0
python-dotenv==1.0.0
textblob
fastapi
uvicorn
salesgpt==0.0.2